alacritty: default TerminalState + safe named color lookup miss
